вынес валидацию в отдельный модуль
