+++ index.html
@@ -105,6 +105,7 @@
           id="address"
           name="address"
           type="text"
+          value = "35.6895, 139.692"
           readonly
           required
           data-pristine-required-message="Обязательное поле">
